+++ src-tauri/src/db/mod.rs
@@ -100,6 +100,7 @@ where
 pub struct ManagerData {
     pub id: i64,
     pub active_game_slug: Option<String>,
+    pub hidden_mods: HashSet<Uuid>,
 }
 
 pub struct ManagedGameData {
@@ -182,17 +183,19 @@ impl Db {
         let conn = self.conn();
 
         let manager = conn
-            .prepare("SELECT id, active_game_slug FROM manager")?
+            .prepare("SELECT id, active_game_slug, hidden_mods FROM manager")?
             .query_row((), |row| {
                 Ok(ManagerData {
                     id: row.get(0)?,
                     active_game_slug: row.get(1)?,
+                    hidden_mods: map_json_row(row, 2)?,
                 })
             })
             .optional()?
             .unwrap_or(ManagerData {
                 id: 1,
                 active_game_slug: None,
+                hidden_mods: HashSet::new(),
             });
 
         let games = conn
@@ -292,9 +295,13 @@ impl Db {
 
     fn _save_manager(&self, tx: &rusqlite::Transaction, manager: &ModManager) -> Result<()> {
         tx.execute(
-            "INSERT OR REPLACE INTO manager (id, active_game_slug)
-            VALUES (?, ?)",
-            params![1, manager.active_game.slug],
+            "INSERT OR REPLACE INTO manager (id, active_game_slug, hidden_mods)
+            VALUES (?, ?, ?)",
+            params![
+                1,
+                manager.active_game.slug,
+                serde_json::to_string(&manager.hidden_mods)?
+            ],
         )?;
 
         Ok(())