+++ src-tauri/src/thunderstore/query.rs
@@ -12,7 +12,7 @@ use super::{
     models::{FrontendMod, FrontendModKind, FrontendVersion, IntoFrontendMod},
 };
 use crate::{
-    profile::{LocalMod, Profile},
+    profile::{LocalMod, ModManager, Profile},
     state::ManagerExt,
     util,
 };
@@ -67,8 +67,7 @@ pub async fn query_loop(app: AppHandle) -> Result<()> {
             if let Some(args) = &thunderstore.current_query {
                 let manager = app.lock_manager();
 
-                let mods =
-                    query_frontend_mods(args, thunderstore.latest(), manager.active_profile());
+                let mods = thunderstore.query_mods(args, &manager);
                 app.emit("mod_query_result", &mods)?;
 
                 if thunderstore.packages_fetched(&app, manager.active_game) {
@@ -239,20 +238,24 @@ impl IntoFrontendMod for LocalMod {
     }
 }
 
-/// Sorts and filters `mods` according to `args` and converts the
-/// results to [`FrontendMod`].
-pub fn query_frontend_mods<T, I>(
-    args: &QueryModsArgs,
-    mods: I,
-    profile: &Profile,
-) -> Vec<FrontendMod>
-where
-    T: Queryable + IntoFrontendMod,
-    I: Iterator<Item = T>,
-{
-    query_mods(args, mods)
+impl Thunderstore {
+    /// Sorts and filters `mods` according to `args` and converts the
+    /// results to [`FrontendMod`].
+    pub(super) fn query_mods(
+        &self,
+        args: &QueryModsArgs,
+        manager: &ModManager,
+    ) -> Vec<FrontendMod> {
+        let profile = manager.active_profile();
+
+        query_mods(
+            args,
+            self.latest()
+                .filter(|borrowed| !manager.hidden_mods.contains(&borrowed.package.uuid)),
+        )
         .map(|m| m.into_frontend(Some(profile)))
         .collect()
+    }
 }
 
 /// Sorts and filters `mods` according to `args`.