+++ src/lib/components/prefs/ApiKeyPref.svelte
@@ -9,11 +9,10 @@
 	import type { Backend } from '$lib/types';
 
 	type Props = {
-		field?: Snippet;
 		backend: Backend;
 	};
 
-	let { field, backend }: Props = $props();
+	let { backend }: Props = $props();
 
 	let hasToken = $state(false);
 
@@ -54,7 +53,5 @@
 		<div class="text-primary-300 group-hover:text-primary-200 truncate">
 			{m[`apiKeyPref_hasToken_${hasToken ? 'has' : 'no'}`]()}
 		</div>
-
-		{@render field?.()}
 	</button>
 </div>