+++ src/lib/components/ui/ModCardList.svelte
@@ -26,9 +26,3 @@
 		</ModCard>
 	{/each}
 </div>
-
-<style>
-	.list {
-		grid-template-columns: repeat(auto-fill, minmax(17rem, 1fr));
-	}
-</style>