+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+### Added
+
+- Context menu options to hide mods from the mod list
+  - This does not prevent them from being installed or shown in the profile mod list
+  - Hidden mods can be unhidden from game-specific manager settings
+
 ### Changed
 
 - Made error message more informative when a profile code couldn't be exported due to upload size limits